refactor: move types to entities, change features folder structure, move api send message function to shared
