+++ client/src/entities/message/types/Message.ts
