--- client/src/features/COPYME/sendMessege/model/apiSendMessage.ts
@@ -1,10 +1,0 @@
-import { axiosInstance, API_URL } from "@/shared/config/axiosInstance";
-import type { ReadyMessage } from "@/features/Chat/sendMessege/model/types/Message";
-
-export const apiSendMessage = async (message: ReadyMessage, room: number) => {
-    const res = await axiosInstance.post(`${API_URL}/send-message`, {
-        message,
-        room,
-    });
-    return res;
-};
