+++ client/src/features/chat-members/model/useGetMembers.ts
@@ -1,7 +1,7 @@
-import { getAllMembers } from "@/shared/api/chat/getMembers";
+import { getAllMembers } from "@/shared/api/member/getMembers";
 import { useSocket } from "@/shared/config/useSocketStore";
 import { useCurrentChatStore } from "@/shared/model/currentChatStore";
-import type { MemberWithName } from "@/entities/chat/model/types/Member";
+import type { MemberWithName } from "@/entities/member/types/Member";
 
 export const useGetMembers = () => {
     const currentChatStore = useCurrentChatStore();
