+++ client/src/shared/api/message/getMessages.ts
@@ -1,4 +1,4 @@
-import type { MessageWithName } from "@/features/Chat/sendMessege/model/types/Message";
+import type { MessageWithName } from "@/entities/message/types/Message";
 import { API_URL, axiosInstance } from "@/shared/config/axiosInstance";
 import type { AxiosResponse } from "axios";
